pyproject.toml
[tool.poetry]
name = "ambernote"
version = "0.1.0"
description = "Note-taking webapp built with Django"
authors = ["Xiangsong Zeng"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
dj-rest-auth = "^3.0.0"
drf-yasg = "^1.21.5"
django-allauth = "^0.52.0"
dj-database-url = "^1.2.0"
django-cache-url = "^3.4.4"
argon2-cffi = "^21.3.0"
pyyaml = "^6.0"
whitenoise = "^6.3.0"
gunicorn = "^20.1.0"
uvicorn = "^0.20.0"
pymysql = "^1.0.2"
mysqlclient = { version = "^2.1.1", optional = true }
psycopg2-binary = { version = "^2.9.5", optional = true }


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"
django-stubs = { extras = ["compatible-mypy"], version = "^1.14.0" }
djangorestframework-stubs = { extras = ["compatible-mypy"], version = "^1.8.0" }

[tool.poetry.extras]
mysqlclient = ["mysqlclient"]
psycopg2 = ["psycopg2-binary"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]


[tool.django-stubs]
django_settings_module = "ambernote.settings.development"
